+++ google2/CHANGELOG.md
@@ -7,7 +7,7 @@ Changed
 - Add `generation` to `GetMetadataResponse`
 - Add `generation` and `metadata` as optional fields for `GoogleStorageService.storeObject`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.3-149b82c"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.3-TRAVIS-REPLACE-ME"`
 
 ## 0.2