+++ google2/CHANGELOG.md
@@ -3,6 +3,10 @@
 This file documents changes to the `workbench-google2` library, including notes on how to upgrade to new versions.
 
 ## 0.3
+
+Added
+- Add `downloadObject`
+
 Changed
 - Add `generation` to `GetMetadataResponse`
 - Add `generation` and `metadata` as optional fields for `GoogleStorageService.storeObject`